Test Ingestion and CCdUpload separately
